% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-registered_cars.R
\docType{data}
\name{registered_cars}
\alias{registered_cars}
\title{Registered vehicle reference data (internal)}
\description{
Internal reference table of registered vehicle makes and models used to
support vehicle make and model matching.
}
\details{
The data are derived from official UK vehicle licensing statistics published
by the Driver and Vehicle Licensing Agency (DVLA).

Source:
\url{https://www.gov.uk/government/statistical-data-sets/vehicle-licensing-statistics-data-tables}
The data have been transformed and reduced for use within this package and
do not represent the full published dataset.
}
\keyword{internal}
